requirements.txt
pycryptodome==3.17
PyInstaller==5.9.0
